Refactor do banco de dados. (#3265)

+++ lib/domain/songbook/use_cases/clear_versions_from_songbook.dart
@@ -6,21 +6,25 @@ import 'package:injectable/injectable.dart';
 import 'package:typed_result/typed_result.dart';
 
 @injectable
-class ClearSongsFromSongbook {
+class ClearVersionsFromSongbook {
   final SongbookRepository _songbookRepository;
   final UserVersionRepository _userVersionRepository;
   final UpdateSongbookPreview _updateSongbookPreview;
 
-  ClearSongsFromSongbook(this._songbookRepository, this._userVersionRepository, this._updateSongbookPreview);
+  ClearVersionsFromSongbook(this._songbookRepository, this._userVersionRepository, this._updateSongbookPreview);
 
   Future<Result<void, RequestError>> call(int songbookId) async {
     final cifrasIdList = (await _userVersionRepository.getUserVersionsFromSongbook(songbookId))
         .map((e) => e.remoteDatabaseId ?? -1)
-        .toList();
-    final apiResult = await _songbookRepository.deleteVersions(songbookId: songbookId, versionsId: cifrasIdList);
-    return apiResult.onSuccess((_) async {
-      await _userVersionRepository.deleteVersionsBySongbookId(songbookId);
-      await _updateSongbookPreview(songbookId);
+        .toList()
+      ..removeWhere((element) => element < 0);
+    return _songbookRepository.deleteVersions(songbookId: songbookId, versionsId: cifrasIdList).then((value) async {
+      if (value.isSuccess) {
+        await _userVersionRepository.deleteVersionsBySongbookId(songbookId);
+        await _updateSongbookPreview(songbookId);
+      }
+
+      return value;
     });
   }
 }
